chatbot_clickhouse/requirements.txt
langchain-mcp-adapters
langgraph 
langchain[anthropic]
streamlit
python-dotenv